refactor(components): move EventEnum to const directory and update import paths
