+++ frontend/src/components/sidebar.tsx
@@ -7,7 +7,7 @@ import SidebarSkeleton from './sidebar-skeleton';
 import UserSettings from './user-settings';
 import { SideBarItem } from './sidebar-item';
 import { Chat } from '@/graphql/type';
-import { EventEnum } from './enum';
+import { EventEnum } from '../const/EventEnum';
 import {
   SidebarContent,
   SidebarGroup,
