+++ frontend/src/const/EventEnum.ts
