+++ frontend/src/const/MessageType.ts
